cloud-run-rag-search/1-apps/apps/rag/frontend/pyproject.toml
[project]
name = "frontend"
version = "0.1.0"
description = "Frontend for RAG application based on Vertex Vector Search."
authors = [{ name = "Google LLC" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "google-cloud-aiplatform>=1.96.0",
    "google-cloud-logging>=3.12.1",
    "google-cloud-storage>=2.16.0",
    "google-genai>=1.16.1",
]

[dependency-groups]
lint = ["mypy>=1.16.0", "ruff>=0.11.13"]

[tool.ruff]
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle
    "F",   # pyflakes
    "W",   # pycodestyle warnings
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "RUF", # ruff specific rules
]
ignore = ["E501", "C901", "RUF001"]

[tool.mypy]
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
check_untyped_defs = true
disallow_subclassing_any = true
warn_incomplete_stub = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_unreachable = true
follow_imports = "silent"
ignore_missing_imports = true
explicit_package_bases = true
disable_error_code = ["misc", "no-untyped-call", "no-any-return"]
